+++ src/callbacks/callback_helpers.jl
@@ -61,7 +61,7 @@ function call_every_n_steps(
         previous_step = Ref(0)
         (u, t, integrator) ->
             (previous_step[] += 1) % n == 0 ||
-                (call_at_end && t == integrator.sol.prob.tspan[2])
+            (call_at_end && t == integrator.sol.prob.tspan[2])
     else
         condition
     end
@@ -167,9 +167,9 @@ frequency. (For instance, reporting something more frequently at the beginning o
 simulation, and less frequency later)
 """
 struct CappedGeometricSeriesSchedule <: AbstractSchedule
-    """GeometricSeriesSchedule(integrator) is true every 2^N iterations or every max_steps"""
+    # GeometricSeriesSchedule(integrator) is true every 2^N iterations or every max_steps
     max_steps::Int
-    """Last step that this returned true"""
+    # Last step that this returned true
     step_last::Base.RefValue{Int}
 
     function CappedGeometricSeriesSchedule(max_steps; step_last = Ref(0))